% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT-models.R
\name{add_constraint_to_model}
\alias{add_constraint_to_model}
\title{Add a constraint to a model}
\usage{
add_constraint_to_model(model, parameter, constraint_type, constraint)
}
\arguments{
\item{model}{The model object for which the the constraint should be added}

\item{parameter}{The name of the parameter for which the constraint will be added}

\item{constraint_type}{The type of constraint to add, one of ("range", "vary", "variable_range") – see details}

\item{constraint}{Either c(min, max) value or c(value, min, max) depending on constraint_type – see details}
}
\value{
A model object
}
\description{
This function adds a constraint to a model. Models can only have one constraint per parameter. Running this function multiple times with the same model and parameter will replace the existing constraint.
}
\section{Details}{

There are 3 types of constraint currently implemented:
\itemize{
\item range: Allow the value of this parameter to vary between the 2 values given.
\item vary: Allow the value of this parameter to vary by this ammount since the previous scan
\item variable_range: Allow the value of this parameter to vary by this ammount since the previous scan, provided it remains within the range given.
}
Values of the constraint parameter are thus:
\itemize{
\item range: c(min, max)
\item vary: value
\item variable_range: c(value, min, max)
}
}

\examples{
add_constraint_to_model(model, "height", "range", c(0, 1))
add_constraint_to_model(model, "height", "vary", 0.5)
add_constraint_to_model(model, "height", "variable_range", c(0.5, 0, 1))
}
