% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT-models.R
\name{add_estimation_function_to_model}
\alias{add_estimation_function_to_model}
\title{Add an estimation function to a model}
\usage{
add_estimation_function_to_model(model, estimation_function)
}
\arguments{
\item{model}{The model object to which the estimation function should be added (see details)}
}
\value{
The model object
}
\description{
This function adds an estimation function to a model.
}
\section{Details}{

The model function recieves the scan number, initial guess and xy data.
It should return a list of parameter estimates (can be a sparse list)
}

\examples{
# estimates the height of a peak based on the intensity at its (guessed) centre
guess_height <- function(n, guess, xy) {
  guess_height <- rev(xy[[2]])[findInterval(guess$centre, rev(xy[[1]]))]
  list(height=guess_height)
}
add_estimation_function_to_model(model, guess_height)
}
