% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_fit_object.R
\name{add_model_to_fit}
\alias{add_model_to_fit}
\title{Add a model to a fit}
\usage{
add_model_to_fit(fit, model)
}
\arguments{
\item{fit}{The fit object to add the model to}

\item{model}{The model object to add}
}
\value{
A fit object
}
\description{
This function adds a model object to a fit. The model must have all constraints etc. added prior to adding it to a fit.
}
\examples{
add_model_to_fit(fit,model)
}
