% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCESS-Baseline.R
\name{analyse_baseline}
\alias{analyse_baseline}
\title{NMR Baseline Analysis}
\usage{
analyse_baseline(data)
}
\arguments{
\item{data}{A data frame of the form [ppm, intensity1, ...]}
}
\value{
None
}
\description{
Analyses the baseline for NMR Spectra graphically. See \code{\link[baseline]{baselineGUI}} for information on additional software dependencies.
}
\examples{
analyse_baseline(data)
}
