% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTIL-complex.R
\name{any_complex}
\alias{any_complex}
\title{Checks whether any values in a dataframe are complex}
\usage{
any_complex(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Checks whether any values in a dataframe are complex
}
