% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_process.R
\name{apkpseudo2d}
\alias{apkpseudo2d}
\title{Autophase NMR Data with constraints for a pseudo 2D dataset (e.g. in-situ echem). Uses (\code{\link[NMR.Utils]{apk}})}
\usage{
apkpseudo2d(spectra, p0_optim_x_range, p1_optim_x_range, pivot, p0_optim_range,
  p1_optim_range)
}
\arguments{
\item{spectra}{(complex) dataframe containing spectra}

\item{p0_optim_x_range}{x range over which to optimise the p0 value (ppm)}

\item{p1_optim_x_range}{x range over which to optimise the p1 value (ppm)}

\item{pivot}{Pivot point for 2nd-order phase correction coefficient (ppm)}

\item{p0_optim_range}{range over which the p0 value can vary}

\item{p1_optim_range}{range over which the p1 value can vary}
}
\value{
A data frame containing the phased NMR data (ppm,intensity1, ...)
}
\description{
This function phases (complex) NMR data.
}
\examples{
apkpseudo2d(data,c(-200,200),c(1000,1300),0,c(-50,50),c(-5,0))
}
