% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr.data.object.R
\name{as.nmr.data.object}
\alias{as.nmr.data.object}
\title{Convert a data.frame to an NMR data object}
\usage{
as.nmr.data.object(x)
}
\arguments{
\item{x}{The object to be converted}
}
\value{
The converted object
}
\description{
Convert a data.frame to an NMR data object
}
\keyword{internal}
