% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr2dinsitu.data.object.R
\name{as.nmr2dinsitu.data.object}
\alias{as.nmr2dinsitu.data.object}
\title{Convert a 2D NMR data object to an insitu NMR data object}
\usage{
as.nmr2dinsitu.data.object(x)
}
\arguments{
\item{x}{The object to be converted}
}
\value{
The converted object
}
\description{
Convert a 2D NMR data object to an insitu NMR data object
}
\keyword{internal}
