% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu.nmr.2Ddata.object.R
\name{associate_echem_with_nmr}
\alias{associate_echem_with_nmr}
\title{Make an insitu NMR data object containing both NMR and Echem data}
\usage{
associate_echem_with_nmr(nmr, echem)
}
\arguments{
\item{nmr}{An nmr.2Ddata.object}

\item{Echem}{The echem data}
}
\value{
An insitu.nmr.2Ddata.object
}
\description{
Make an insitu NMR data object containing both NMR and Echem data
}
