% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_plot.R
\name{color.scale.nmr}
\alias{color.scale.nmr}
\title{Colour NMR Data}
\usage{
color.scale.nmr(x, cs1 = c(0, 1), cs2 = c(0, 1), cs3 = c(0, 1),
  alpha = 1, extremes = NA, na.color = NA, xrange = NULL,
  color.spec = "rgb")
}
\description{
This function is a workaround for an issue with \code{\link[plotrix]{color.scale}}.
}
\keyword{internal}

