% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCESS-Baseline.R
\name{correct_baseline}
\alias{correct_baseline}
\title{NMR Baseline correction}
\usage{
correct_baseline(data, ...)
}
\arguments{
\item{data}{A data frame of the form [ppm, intensity1, ...]}

\item{...}{Additional parameters are passed to the baseline function (\code{\link[baseline]{baseline}})}
}
\value{
An identical data frame containing the baseline corrected data
}
\description{
Corrects the baseline for NMR Spectra
}
\examples{
correct_baseline(data, method="modpolyfit", degree=1)
}
