% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_objects.R
\name{nmr.data.object}
\alias{nmr.data.object}
\alias{nmr2d.data.object}
\alias{nmr2dinsitu.data.object}
\alias{as.nmr.data.object}
\alias{as.nmr2d.data.object}
\alias{as.nmr2dinsitu.data.object}
\alias{is.nmr.data.object}
\alias{is.nmr2d.data.object}
\alias{is.nmr2dinsitu.data.object}
\title{NMR data objects}
\usage{
nmr.data.object(...)

nmr2d.data.object(...)

nmr2dinsitu.data.object(...)

as.nmr.data.object(x)

as.nmr2d.data.object(x)

as.nmr2dinsitu.data.object(x)

is.nmr.data.object(x)

is.nmr2d.data.object(x)

is.nmr2dinsitu.data.object(x)
}
\arguments{
\item{x}{The object}
}
\keyword{internal}
