% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPORT-nmr-guess.R
\name{determine_1d_file_type}
\alias{determine_1d_file_type}
\title{Determine NMR file type}
\usage{
determine_1d_file_type(file)
}
\arguments{
\item{file}{File to read}
}
\value{
An integer representing the file type
}
\description{
This function attempts to determine the type of a bruker text file
}
\examples{
determine_1d_file_type(file)
}
\keyword{internal}
