% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_import.R
\name{determine_dimensions_from_text_file}
\alias{determine_dimensions_from_text_file}
\title{Read NMR Data Dimensions}
\usage{
determine_dimensions_from_text_file(file)
}
\arguments{
\item{file}{Path to txt file saved using topspin's File->Save->Save visible region as text file (with or without imaginary)}
}
\value{
A number indicating the dimensions
}
\description{
This function attempts to determine the dimension of a bruker text file by examining the preamble.
}
\examples{
determine_dimensions_from_text_file("/path/to/file.txt")
}
\keyword{internal}
