% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echem_import.R
\name{dummy.echem}
\alias{dummy.echem}
\title{Produce dummy echem data}
\usage{
dummy.echem()
}
\value{
A data frame containing the dummy echem data
}
\description{
This function is used to produce dummy echem data.
}
\examples{
dummy.echem()
}

