% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echem_import.R
\name{echem.combine}
\alias{echem.combine}
\title{Combine echem experiments}
\usage{
echem.combine(echem_1, echem_2, time_offset = 0)
}
\value{
A data frame containing the echem data
}
\description{
This function is used to combine two echem experiments, adjusting the time of the second experiment to follow the 1st.
}
\examples{
echem.combine(echem_1,echem_2)
}

