% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT-fit_object.R
\name{exclude_ppm_range_from_fit}
\alias{exclude_ppm_range_from_fit}
\title{Exclude a ppm range from the fit}
\usage{
exclude_ppm_range_from_fit(fit, range)
}
\arguments{
\item{fit}{The fit object from which to exclude the ppm range.}

\item{range}{The range to exclude c(start,end) -- start < end}
}
\value{
The fit object
}
\description{
This function excludes a ppm range from the fit.
}
