% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT-fit_object.R
\name{export.nmr.fit.object}
\alias{export.nmr.fit.object}
\title{Export fit results}
\usage{
\method{export}{nmr.fit.object}(fit, filename)
}
\arguments{
\item{fit}{The fit object}

\item{filename}{The path to the file (ending in .csv)}
}
\value{
None
}
\description{
This function is used to export the fit results as a csv file.
}
\examples{
export(fit,"/path/to/file.csv")
}
