% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA-export.R
\name{export.nmr2dinsitu.data.object}
\alias{export.nmr2dinsitu.data.object}
\title{Export insitu NMR data}
\usage{
\method{export}{nmr2dinsitu.data.object}(data, path)
}
\description{
Export insitu NMR data
}
