% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_echem}
\alias{export_echem}
\title{Export echem data}
\usage{
export_echem(echem, filename)
}
\arguments{
\item{echem}{The echem data object}

\item{filename}{The path to the file (ending in .csv)}
}
\value{
None
}
\description{
This function is used to export echem data as a csv file.
}
\examples{
export_echem(echem,"/path/to/file.csv")
}

