% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_fit}
\alias{export_fit}
\title{Export fit results}
\usage{
export_fit(fit, filename)
}
\arguments{
\item{fit}{The fit object}

\item{filename}{The path to the file (ending in .csv)}
}
\value{
None
}
\description{
This function is used to export the fit results as a csv file.
}
\examples{
export_fit(fit,"/path/to/file.csv")
}

