% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_nmr}
\alias{export_nmr}
\title{Export NMR Data}
\usage{
export_nmr(data, filename, scan_number = NA)
}
\arguments{
\item{data}{The data object}

\item{filename}{The path to the file (ending in .csv)}

\item{scan_number}{For 2D data, the scan number may be (optionally) specified}
}
\value{
None
}
\description{
This function is used to export 1D or 2D NMR data as a csv file.
}
\examples{
export_nmr(data,"/path/to/file.csv")

export_nmr(data,"/path/to/file.csv",scan_number=1)
}

