% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCESS-reduce_scans.R
\name{find_last_scan}
\alias{find_last_scan}
\title{Finds the number of the last (non-empty) scan}
\usage{
find_last_scan(spectra, tol = 0.01)
}
\arguments{
\item{spectra}{A dataframe of NMR data}

\item{tol}{If all the intensities are equal to within this tolerance the scan is considered empty}
}
\value{
The number of the last scan
}
\description{
Finds the number of the last (non-empty) scan
}
\examples{
find_last_scan(spectra)
}
