% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_leastSquares.R
\name{fit_initial_brute_force}
\alias{fit_initial_brute_force}
\title{Brute force the starting parameter}
\usage{
fit_initial_brute_force(models, x, y, formula, upper, lower,
  grid_points = 100)
}
\arguments{
\item{models}{The models to use}

\item{x}{The x data to fit}

\item{y}{The y data to fit}

\item{formula}{The nls formula}

\item{upper}{The upper bounds}

\item{lower}{The lower bounds}

\item{grid_points}{The number of points in the grid to sample}
}
\value{
The best coefficients
}
\description{
This function runs many fits to sample the landscape for local minima.
}
\examples{
fit_initial_brute_force(models,x,y,formula,upper,lower)
}
\keyword{internal}
