% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_leastSquares.R
\name{fit_single_scan}
\alias{fit_single_scan}
\title{Run a fit on a single scan}
\usage{
fit_single_scan(fit, data)
}
\arguments{
\item{fit}{The fit object to use}

\item{data}{The data to fit}
}
\value{
The fit object
}
\description{
This function fits a single scan using the models defined previously.
}
\examples{
fit_single_scan(fit, data)
}
\keyword{internal}

