% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_fit_object.R
\name{fix_unconstrained_parameters}
\alias{fix_unconstrained_parameters}
\title{Fixes the uncronstrained parameters for every model}
\usage{
fix_unconstrained_parameters(fit)
}
\arguments{
\item{fit}{The fit object to use}
}
\value{
The fit object
}
\description{
This function modifies the function for each model, adding default values for the parameters not to be fitted.
}
\examples{
fix_unconstrained_parameters(fit)
}
\keyword{internal}

