% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_fit_object.R
\name{get_fit_function}
\alias{get_fit_function}
\title{Generate fit function}
\usage{
get_fit_function(models)
}
\arguments{
\item{models}{List of models to use}
}
\value{
The fit function
}
\description{
This function creates a function representing the overall fit
}
\examples{
func=get_fit_function(models)
}
\keyword{internal}
