% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-BASELINE.R, R/GUI-IMPORT.R, R/GUI-PHASE.R,
%   R/GUI-PLOTTING.R, R/GUI-WIZARD.R
\name{interactiveBaseline}
\alias{interactiveBaseline}
\alias{interactiveImport}
\alias{interactivePhase}
\alias{interactivePlotting}
\alias{insitu_gui}
\title{Wizard for processing insitu data}
\usage{
interactiveBaseline(nmr)

interactiveImport()

interactivePhase(nmr)

interactivePlotting(nmr)

insitu_gui(nmr)
}
\arguments{
\item{nmr}{The NMR data}
}
\description{
Use \code{insitu_gui} for the main wizard, or access an individual tab with its respective command.
You do not need to provide the \code{nmr} argument to \code{insitu_gui}, doing so will disable the import tab.\cr\cr
By default logs will be output to the console during operation, you can adjust the level of detail using the command:\cr
\code{jms.classes::set_persistent_setting('NMR-GUI-LOGLEVEL', 'DEBUG')}\cr
Replacing \code{'DEBUG'} with \code{'INFO'}, \code{'WARN'}, or \code{'ERROR'}.
You can disable logging entirely using \code{'FALSE'}. R may need to be restarted before this command will take effect.
}
\examples{
insitu_gui()
}
