% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_leastSquares.R
\name{integrate_last_scan}
\alias{integrate_last_scan}
\title{Integrate a fitted function}
\usage{
integrate_last_scan(fit, data)
}
\arguments{
\item{fit}{The fit object to use}

\item{data}{The data object}
}
\value{
The fit object
}
\description{
This function fits the last fitted scan using the models defined previously. Adds the result into fit$result.
}
\examples{
integrate_last_scan(fit)
}
\keyword{internal}
