% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-BASELINE.R
\name{interactiveBaseline}
\alias{interactiveBaseline}
\title{Creates a GUI for correcting the baseline of NMR data}
\usage{
interactiveBaseline(nmr)
}
\arguments{
\item{nmr}{The NMR data}
}
\description{
Creates a GUI for correcting the baseline of NMR data
}
