% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCESS-Interactive_phase.R
\name{interactivePhase}
\alias{interactivePhase}
\title{Creates a GUI for phasing data}
\usage{
interactivePhase(nmr, allow_real = FALSE)
}
\arguments{
\item{nmr}{Complex NMR data to be phased}

\item{allow_real}{Allow the error message that is shown if real data are provided to be dismissed}
}
\description{
Creates a GUI for phasing data
}
