% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-PLOTTING.R
\name{interactivePlotting}
\alias{interactivePlotting}
\title{Creates a GUI for generating plots of insitu data}
\usage{
interactivePlotting(nmr)
}
\arguments{
\item{nmr}{The NMR data}
}
\description{
Creates a GUI for generating plots of insitu data
}
