% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT-nmr.data.object.R
\name{iplot.nmr.data.object}
\alias{iplot.nmr.data.object}
\title{Plot ex situ NMR data}
\usage{
\method{iplot}{nmr.data.object}(..., axes = 1, offset = NULL,
  xlim = NULL, ylim = NULL, y2lim = NULL, xlab = NULL,
  ylab = NULL, y2lab = NULL, col = NULL, lwd = NULL, pch = NULL,
  labels = NULL, group = NULL)
}
\arguments{
\item{...}{Additional parameters passed to \code{\link[jms.classes]{iplot.jms.data.object}}}

\item{x}{An nmr.data.object}
}
\description{
This function is used to plot ex situ NMR data.
}
