% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu.nmr.2Ddata.object.R
\name{is.insitu.nmr.2Ddata.object}
\alias{is.insitu.nmr.2Ddata.object}
\title{Check if an object is an insitu NMR data object}
\usage{
is.insitu.nmr.2Ddata.object(x)
}
\arguments{
\item{x}{The object to be tested}
}
\value{
TRUE / FALSE
}
\description{
Check if an object is an insitu NMR data object
}
\keyword{internal}
