% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.table.nmr.R
\name{is.nmr.data.object}
\alias{is.nmr.data.object}
\title{Check if an object is an NMR data object}
\usage{
is.nmr.data.object(x)
}
\arguments{
\item{x}{The object to be tested}
}
\value{
TRUE / FALSE
}
\description{
Check if an object is an NMR data object
}
\keyword{internal}
