% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_jms.reexports.R
\name{reexports}
\alias{reexports}
\alias{iplot}
\alias{export}
\alias{xlab}
\alias{ylab}
\alias{y2lab}
\alias{xcol}
\alias{ycol}
\alias{y2col}
\alias{make_background}
\alias{undersample}
\title{Reexported functions from jms.classes}
\description{
Reexported functions from jms.classes
}
\keyword{internal}
