% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echem_import.R
\name{load.biologic}
\alias{load.biologic}
\title{Read Echem Data for Biologic Cycler}
\usage{
load.biologic(DL_file)
}
\arguments{
\item{DL_file}{Path to data file}
}
\value{
A data frame containing the echem data
}
\description{
This function is used to import echem data from a biologic mpt file.
}
\examples{
load.biologic("/path/to/file.mpt")
}
\keyword{internal}

