% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echem_import.R
\name{load.land}
\alias{load.land}
\title{Read Echem Data for land (lahne) Cycler}
\usage{
load.land(DL_file)
}
\arguments{
\item{DL_file}{Path to data file}
}
\value{
A data frame containing the echem data
}
\description{
This function is used to import echem data from a land txt file.
}
\examples{
load.land("/path/to/file.txt")
}
\keyword{internal}

