% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{load_fit}
\alias{load_fit}
\title{Load fit}
\usage{
load_fit(filename)
}
\arguments{
\item{filename}{The path to the file (ending in .R)}
}
\value{
None
}
\description{
This function is used to load a fit saved using (\code{\link[NMR.Utils]{save_fit}}).
}
\examples{
fit = load_fit("/path/to/file.R")
}

