% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_or_install}
\alias{load_or_install}
\title{Load packages}
\usage{
load_or_install(x)
}
\arguments{
\item{x}{A (complex) dataframe of NMR data}
}
\value{
None
}
\description{
This function will load a package, or install it if it isn't present. (Probably against all sorts of programming guidelines, but makes everything easier to install and run...)
}
\examples{
load_or_install("baseline")
}
\keyword{internal}

