% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu.nmr.2Ddata.object.R
\name{make.insitu.nmr.2Ddata.object}
\alias{make.insitu.nmr.2Ddata.object}
\title{Convert a 2D NMR data object to an insitu NMR data object}
\usage{
make.insitu.nmr.2Ddata.object(x)
}
\arguments{
\item{x}{The object to be converted}
}
\value{
The converted object
}
\description{
Convert a 2D NMR data object to an insitu NMR data object
}
\keyword{internal}
