% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr.data.object.R
\name{make.nmr.2Ddata.object}
\alias{make.nmr.2Ddata.object}
\title{Convert an NMR data object to a 2D NMR data object}
\usage{
make.nmr.2Ddata.object(x)
}
\arguments{
\item{x}{The object to be converted}
}
\value{
The converted object
}
\description{
Convert an NMR data object to a 2D NMR data object
}
\keyword{internal}
