% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.table.nmr.R
\name{make.nmr.data.object}
\alias{make.nmr.data.object}
\title{Convert a data.frame to an NMR data object}
\usage{
make.nmr.data.object(x)
}
\arguments{
\item{x}{The object to be converted}
}
\value{
The converted object
}
\description{
Convert a data.frame to an NMR data object
}
\keyword{internal}
