% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_fit_object.R
\name{make_dynamic_fit}
\alias{make_dynamic_fit}
\title{Make dynamic fit}
\usage{
make_dynamic_fit(fit, scans, update_guess_for_non_fixed = TRUE)
}
\arguments{
\item{fit}{The fit object to make dynamic}

\item{scans}{Scans containing only a single peak in the region of interest (one per model peak)}

\item{update_guess_for_non_fixed}{Use the automatically determined parameters for non-fixed parameters (as well as for the fixed ones)}
}
\value{
The fit object
}
\description{
This function makes the fit run as a dynamic fit. For this to work effectively, there must be one scan per model peak that contains ONLY this peak in the region of interest (no overlap). Otherwise the model will have to be fully defined manually.
}

