% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_fit_object.R
\name{make_fitted_models}
\alias{make_fitted_models}
\title{Generate a list of model functions that only take an x parameter}
\usage{
make_fitted_models(models, result_line)
}
\arguments{
\item{models}{The models list}

\item{result_line}{A row from the results table}
}
\value{
list(c(model_index, fitted_model),...)
}
\description{
This function converts a list of models to a list of functions + model indexes.
}
\examples{
make_fitted_models(models,result_line)
}
\keyword{internal}
