% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_process.r
\name{makeReal}
\alias{makeReal}
\title{Removes imaginary component from NMR data}
\usage{
makeReal(spectra)
}
\arguments{
\item{spectra}{A (complex) dataframe of NMR data}
}
\value{
A data frame containing the real NMR data (ppm,intensity1, ...)
}
\description{
Removes imaginary component from NMR data
}
\examples{
makeReal(data)
}
