% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_fit_object.R
\name{new_fit}
\alias{new_fit}
\title{Create a new fit}
\usage{
new_fit(integration_range = c(-Inf, Inf), noise_height = 0,
  maxEvaluations = 10000, maxIterations = 1000,
  use_previous_as_guess = TRUE, brute_force = FALSE,
  brute_force_method = "short", brute_force_grid_size = 10)
}
\arguments{
\item{integration_range}{The range over which to perform the integration c(min,max)}

\item{noise_height}{Approximate magnitude of the noise in the spectrum (probably best as 0)}

\item{maxEvaluations}{Max number of calls to model function}

\item{maxIterations}{Max number of iterations during fitting}

\item{use_previous_as_guess}{Uses the previous fit as a guess for the next fit (rather than initial guess)}

\item{brute_force}{Attempt to brute force the starting parameters}

\item{brute_force_method}{'short' or 'full'. Short only tries to fit using all of the models, full will also try every combination of models.}

\item{brute_force_grid_size}{Number of parameters to try for every constraint during the brute force.}
}
\value{
A fit object
}
\description{
This function creates a fit that can contain models.
}
\examples{
new_fit()
}
