% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_models.R
\name{new_model}
\alias{new_model}
\title{Create a new model}
\usage{
new_model(type, ...)
}
\arguments{
\item{type}{A function which defines the curve to fit (\code{\link[NMR.Utils]{pseudoVoigt}} is the only predefined model function)}

\item{...}{Initial guesses for the parameters of the function}
}
\value{
A model object
}
\description{
This function creates a model to be fitted.
}
\examples{
new_model(pseudoVoigt, height=1e9,centre=-9.95,hwhm=2.1,shape=1)
}
