% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.table.nmr.R
\name{nmr.data.object}
\alias{nmr.data.object}
\title{Make a new NMR data object}
\usage{
nmr.data.object(...)
}
\arguments{
\item{...}{parameters are passed to \code{\link{data.frame}}}
}
\value{
An NMR data object
}
\description{
Make a new NMR data object
}
\keyword{internal}
