% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_process.r
\name{noATMoffsets}
\alias{noATMoffsets}
\title{Generates time offsets if spectra were acquired without ATM}
\usage{
noATMoffsets(spectra, scan_length, initial_offset = 0)
}
\arguments{
\item{spectra}{A dataframe of NMR data}

\item{scan_length}{Length of a single scan (hours)}

\item{initial_offset}{Offset of 1st scan (hours)}
}
\value{
A vector of offsets
}
\description{
Generates time offsets if spectra were acquired without ATM
}
\examples{
noATMoffsets(spectra,0.5)
}
