% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_process.R
\name{phase}
\alias{phase}
\title{Phase NMR Data}
\usage{
phase(data, p0, p1, pivot)
}
\arguments{
\item{data}{A (complex) dataframe of NMR data}

\item{p0}{0th-order phase correction coefficient}

\item{p1}{1st-order phase correction coefficient}

\item{pivot}{Pivot point for 2nd-order phase correction coefficient (ppm)}
}
\value{
A data frame containing the phased NMR data (ppm,intensity1, ...)
}
\description{
This function phases (complex) NMR data.
}
\examples{
phase(data,45,0,0)
}
