% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu_plot.R
\name{plot.insitu.nmr.2Ddata.object}
\alias{plot.insitu.nmr.2Ddata.object}
\title{Plot electrochemisty data (vertically)}
\usage{
\method{plot}{insitu.nmr.2Ddata.object}(nmr, widths = c(3, 1), ...)
}
\arguments{
\item{nmr}{Data frame of nmr data}

\item{...}{Further parameters will be passed to \code{\link[NMR.Utils]{plot.nmr.2Ddata.object}} or \code{\link[NMR.Utils]{plotechem_vertical}}}

\item{echem}{Data frame of echem data}
}
\description{
This function plots in-situ echem data.
}
\examples{
plot(data,xrange=c(1500,-500), yrange='auto', plot_offset=100000, plot.colour.ranges=list(c(210,900)), plot.colour.yranges=list(c(50000,200000)),V_range=c(0,2),alignment=align)
}
