% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT-nmr.data.object.R
\name{plot.nmr.data.object}
\alias{plot.nmr.data.object}
\title{Plot ex situ NMR data}
\usage{
\method{plot}{nmr.data.object}(x, ..., axes = 1, xlim = NULL)
}
\arguments{
\item{x}{An nmr.data.object}

\item{...}{Additional parameters passed to \code{\link[jms.classes]{plot.jms.data.object}}}
}
\description{
This function is used to plot ex situ NMR data.
}
