% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nmr.fit.object.R
\name{plot.nmr.fit.object}
\alias{plot.nmr.fit.object}
\title{Plot the integrated intensities}
\usage{
\method{plot}{nmr.fit.object}(fit, showErrors = TRUE, NAasZero = FALSE,
  showLegend = TRUE, pch = 4, fit_col = "red",
  deconv_cols = c("forestgreen", "blue", "magenta", "cyan", "orange",
  "brown"), error_alpha = 0.3, plot_total_integral = TRUE)
}
\arguments{
\item{fit}{The fit object}

\item{showErrors}{Should the errors be displayed (as shaded regions) (Default TRUE)}

\item{NAasZero}{Display Models not included in the fit with 0 area and 0 error? (Default FALSE)}

\item{showLegend}{Show a legend above the plot? (Default TRUE)}

\item{pch}{The character to use for points (3 = + , 4 = x , 16 = filled circle etc.) (see \code{\link[graphics]{points}})}

\item{fit_col}{Colour for the total integral}

\item{deconv_cols}{Colours for the deconvolution integrals}

\item{error_alpha}{Factor to change the alpha when plotting errors}

\item{plot_total_integral}{TRUE / FALSE should plot the total integral}
}
\value{
None
}
\description{
This function plots the integrated intensities of the fit and the deconvolution for all frames in the fit.
}
\examples{
plot(fit)
}
