% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_plot.R
\name{plot.nmr2d.data.object}
\alias{plot.nmr2d.data.object}
\title{Plot 2D NMR data}
\usage{
\method{plot}{nmr2d.data.object}(nmrdata, xlim = NA, ylim = NA,
  plot_offset = "auto", plot.colour = TRUE, plot.colour.ranges = NA,
  plot.colour.yranges = "auto", y_trunc = NA, col = "Black",
  lwd = par("lwd"), shade_under = FALSE, shade_col = "grey",
  y_trunc_x_points = c(), y_trunc_amp_div = 200,
  y_trunc_label_offset_factor = 20, y_trunc_sin_period = 5,
  y_trunc_labels = c(), y_trunc_text_col = "grey",
  y_trunc_line_col = "grey", y_trunc_lwd = 2, y_trunc_cex = 1, axes = 1,
  show_RH_Tick = TRUE, show_LH_Tick = TRUE, xaxismline = -0.8,
  xaxislabelmline = 1.1, col_na = "black", colour_scheme = c("blue",
  "green", "yellow", "magenta", "red"), colour.legend = FALSE,
  colour.legend.show.zero = TRUE, show_ppm_label = TRUE)
}
\arguments{
\item{nmrdata}{Data frame of (real) NMR data (ppm, intensity1, ...)}

\item{xlim}{x range of data to plot (ppm) -- should be in form c(upper, lower) or spectrum will be reversed (or 'auto')}

\item{ylim}{Y range of data to plot (or 'auto')}

\item{plot_offset}{offset between subsequent plots (or 'auto')}

\item{plot.colour}{(TRUE / FALSE) Should the plot be coloured?}

\item{plot.colour.ranges}{list of x ranges to colour (can be NA to cover entire range)}

\item{y_trunc}{Truncate y values that are higher than this (or NA for auto)}

\item{col}{Line colour for plot}

\item{lwd}{Line width for plot}

\item{shade_under}{TRUE / FALSE shade under the plot}

\item{shade_col}{colour to shade plot}

\item{y_trunc_x_points}{X-ranges to draw truncation lines over}

\item{y_trunc_amp_div}{Factor to divide amplitude of truncation line}

\item{y_trunc_label_offset_factor}{Factor to offset truncation line label by}

\item{y_trunc_sin_period}{Period of sine wave for truncation line}

\item{y_trunc_labels}{List of labels for truncation line}

\item{y_trunc_text_col}{Colour of text for truncation line}

\item{y_trunc_line_col}{Colour of truncation line}

\item{y_trunc_lwd}{Line width of truncation line}

\item{y_trunc_cex}{cex of truncation label}

\item{show_RH_Tick}{(TRUE / FALSE) Show last tick on RHS}

\item{show_LH_Tick}{(TRUE / FALSE) Show last tick on LHS}

\item{xaxismline}{The margin line on which to draw the x-axis}

\item{xaxislabelmline}{The margin line on which to draw the x-axis label}

\item{col_na}{Colour of values outside of colouring y-range}

\item{colour_scheme}{If plot.colour is TRUE, each colour y-range (as specified in plot.colour.ylims) will be split into a number of equal ranges based on the number of colours specified in this vector. Each range will be assigned to one of the colours, starting from the smallest y-value.}

\item{colour.legend}{(TRUE / FALSE) Show a basic legend of the chosen colour_scheme}

\item{colour.legend.show.zero}{(TRUE / FALSE) If the colour range extends beyond zero, indicate this value in the legend (WARNING: Only valid if there is a single colour scaling range)}

\item{show_ppm_label}{(TRUE / FALSE) Should plot ppm axis label}

\item{xTickInterval}{Tick interval for x axis (or 'auto')}

\item{xMinorTickInterval}{Minor tick interval for x axis (or 'auto')}

\item{plot.colour.ylims}{list of y ranges to scale colours over (can be 'auto')}
}
\value{
List of alignment_parameters
}
\description{
This function plots 2D NMR Data.
}
\examples{
plot(data)
}
