% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT-nmr2dinsitu.data.object.R
\name{plot.nmr2dinsitu.data.object}
\alias{plot.nmr2dinsitu.data.object}
\title{Plot In situ NMR data}
\usage{
\method{plot}{nmr2dinsitu.data.object}(nmr, use.default.layout = TRUE,
  separation = 0.1, separation.pos = c(4, 2), nmr.func = NULL,
  echem.func = NULL, ...)
}
\arguments{
\item{nmr}{In situ nmr data}

\item{use.default.layout}{If TRUE, a side-by-side 3:1 layout will be used for NMR-Echem.
If FALSE, \code{\link[graphics]{layout}} must be called prior to calling \code{plot}
Plot #1 will be used for NMR, #2 for echem}

\item{separation}{Gap between plots (in inches)}

\item{separation.pos}{Where to put the separation -- vector c(nmr, echem)
where nmr,echem are one of 1=bottom, 2=left, 3=top, 4=right}

\item{nmr.func}{Function to plot NMR data (default \code{\link[NMR.Utils]{plot.nmr2d.data.object}})}

\item{echem.func}{Function to plot echem data (default \code{\link[NMR.Utils]{plot_echem_vertical}})}

\item{...}{Further parameters will be passed to the plotting methods}
}
\value{
Returns a function that will change the active plot to the NMR plot without disrupting the current layout
(the echem plot is active by default when the plot function returns).
}
\description{
This function plots in situ NMR data.
}
\details{
The echem function will always be passed the echem data restricted to the NMR scans plotted as its first argument,
and a ylim parameter which will align its data as another argument. Any further named arguments provided via \code{...}
are passed through. \cr\cr
The NMR function will always be passed the NMR data as its first argument. Any further named arguments provided via \code{...}
are passed through. It is expected to return a numeric vector of the y values corresponding to an intensity of 0 for the
first and last scans.
}
\examples{
plot(data)

set_nmr <- plot(data)

# Add some annotations to the echem plot
time = c(1, 2)
V = Plotting.Utils::nearest(echem, Test_Time.s.=time*3600)$Voltage.V
points(V, time, col=c('red', 'blue'))

# Activate the nmr plot
set_nmr()
# Add annotations to the NMR plot
points(0, 10000, col='green')
}
