% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animations.R
\name{plot_animation_frame}
\alias{plot_animation_frame}
\title{Plot a raw frame for a fit animation}
\usage{
plot_animation_frame(fit, data, echem, frame_number, xrange, V_range,
  yrange = "auto", include_fit = TRUE, data_col = "black", data_lwd = 1,
  cols = c("forestgreen", "blue", "magenta", "cyan", "orange", "brown"),
  lwd = 1, lty = "dashed", fit_col = "red", fit_lwd = 2,
  xTickInterval = 50)
}
\arguments{
\item{fit}{The fit object}

\item{data}{The data object}

\item{echem}{The echem object}

\item{frame_number}{Which frame to plot}

\item{xrange}{x range of data to plot}

\item{yrange}{Y range of data to plot (or 'auto')}

\item{include_fit}{include the deconvolution etc. in the plot}

\item{data_col}{Colour for the data}

\item{data_lwd}{Line width for the data}

\item{cols}{Colour for each deconvolution}

\item{lwd}{Line width for each deconvolution}

\item{lty}{Line type for each deconvolution}

\item{fit_col}{Colour for the fit}

\item{fit_lwd}{Line width for the fit}

\item{xTickInterval}{Tick interval for x axis}
}
\value{
None
}
\description{
This function plots a single frame from the fit.
}
\examples{
plot_animation_frame(fit,data,echem,100,c(200,-200),c(0,2))
}
\keyword{internal}
