% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echem_plot.R
\name{plot_echem_horizontal}
\alias{plot_echem_horizontal}
\title{Plot electrochemisty data (horizontally)}
\usage{
plot_echem_horizontal(data, V_range, alignment = NA, show_axes = TRUE,
  V_tick_interval = 0.25, time_tick_interval = 5, time_ticks_dps = 0,
  mar = 2)
}
\arguments{
\item{data}{Data frame of echem data}

\item{V_range}{Range of potential (Voltage) to plot}

\item{alignment}{Alignment parameters}

\item{show_axes}{(TRUE / FALSE) Should plot axes & labels}

\item{V_tick_interval}{Tick interval for the potential axis}

\item{time_tick_interval}{Tick interval for the time axis}
}
\description{
This function plots in-situ echem data.
}
\examples{
plot_echem_horizontal(data,V_range)
}
