% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT-1d-legacy.R
\name{plot_ex_situ_nmr}
\alias{plot_ex_situ_nmr}
\title{Plot ex situ NMR data}
\usage{
plot_ex_situ_nmr(data, names = c(""), plot.cols = NA,
  .interactive_xlim = NULL, .interactive_ylim = NULL, ...)
}
\arguments{
\item{data}{A list of data to plot (see \code{\link{read.NMR}})}

\item{names}{Names of the data to go into the legend (defaults to filenames)}

\item{plot.cols}{list of colours for plots (defaults to automatically chosing)}
}
\value{
A list of plotting options to be displayed to the user (invisibly)
}
\description{
WARNING: LEGACY FUNCTION, USE plot(...) INSTEAD
This function is used to plot ex situ NMR data. A grid of plots is produced per nucleus present in the data list.
}
\concept{Ploting Methods}
