% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ex_situ.R
\name{plot_ex_situ_nmr_files}
\alias{plot_ex_situ_nmr_files}
\title{Read and plot ex situ NMR data}
\usage{
plot_ex_situ_nmr_files(files, acqus_dirs = c(), masses = c(), ...)
}
\arguments{
\item{files}{The Paths to the data files}

\item{acqus_dirs}{The Paths to the directories containing acqus files}

\item{...}{Additional parameters passed to \code{\link{plot_ex_situ_nmr}}}
}
\description{
WARNING: LEGACY FUNCTION, USE plot(...) INSTEAD
This function is used to read and plot ex situ NMR data.
}
