% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT-fits-legacy.R
\name{plot_examples}
\alias{plot_examples}
\title{Plot examples of fitted frames}
\usage{
plot_examples(fit, data, scans, plot_offset, yrange = "auto",
  xrange = NA, cols = c("forestgreen", "blue", "magenta", "cyan",
  "orange", "brown"), fit_col = "red", data_col = "black",
  fit_lwd = 2, lwd = 1, data_lwd = 1, lty = "dashed",
  xTickInterval = 50, show_axes = TRUE)
}
\arguments{
\item{fit}{The fit object}

\item{data}{The data object}

\item{scans}{Which scans to include in the plot}

\item{plot_offset}{offset between subsequent plots}

\item{yrange}{Y range of data to plot (or 'auto')}

\item{xrange}{X range of data to plot (default is integration range)}

\item{cols}{Colour for each deconvolution}

\item{fit_col}{Colour for the fit}

\item{data_col}{Colour for the data}

\item{fit_lwd}{Line width for the fit}

\item{lwd}{Line width for each deconvolution}

\item{data_lwd}{Line width for the data}

\item{lty}{Line type for each deconvolution}

\item{xTickInterval}{Tick interval for x axis}

\item{show_axes}{(TRUE / FALSE) Should plot axes & labels}
}
\value{
None
}
\description{
This function plots the deconvolution for specified frames from the fit.
}
\examples{
plot_examples(fit,data,scans=seq(1:20,5),plot_offset=100000)
}
