% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT-fits-legacy.R
\name{plot_fit}
\alias{plot_fit}
\title{Produce some summary plots}
\usage{
plot_fit(fit, data, echem)
}
\arguments{
\item{fit}{The fit object}

\item{data}{The data object}

\item{echem}{The echem object}
}
\value{
None
}
\description{
This function plots some summary plots to show the fit.
}
\examples{
plot_fit(fit,data,echem)
}
