% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_plots.R
\name{plot_integrals}
\alias{plot_integrals}
\title{Plot the integrated intensities}
\usage{
plot_integrals(fit, pch = 4, fit_col = "red",
  deconv_cols = c("forestgreen", "blue", "magenta", "cyan", "orange",
  "brown"), error_alpha = 0.5, plot_total_integral = TRUE, mar = 2,
  show_axes = TRUE)
}
\arguments{
\item{fit}{The fit object}

\item{pch}{The character to use for points (3 = + , 4 = x , 16 = filled circle etc.) (see \code{\link[graphics]{points}})}

\item{fit_col}{Colour for the total integral}

\item{deconv_cols}{Colours for the deconvolution integrals}

\item{error_alpha}{Factor to change the alpha when plotting errors}

\item{plot_total_integral}{TRUE / FALSE should plot the total integral}

\item{mar}{side to show axis on}

\item{show_axes}{(TRUE / FALSE) Should plot axes & labels}
}
\value{
None
}
\description{
This function plots the integrated intensities of the fit and the deconvolution for all frames in the fit.
}
\examples{
plot_integrals(fit)
}

