% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_plots.R
\name{plot_parameter}
\alias{plot_parameter}
\title{Plot a fitted parameter}
\usage{
plot_parameter(fit, parameter_name, pch = 4, cols = c("forestgreen", "blue",
  "magenta", "cyan", "orange", "brown"), show_axes = TRUE)
}
\arguments{
\item{fit}{The fit object}

\item{parameter_name}{The name of the parameter to plot}

\item{pch}{The character to use for points (3 = + , 4 = x , 16 = filled circle etc.) (see \code{\link[graphics]{points}})}

\item{cols}{Colours for the parameters}

\item{show_axes}{(TRUE / FALSE) Should plot axes & labels}
}
\value{
None
}
\description{
This function plots a parameter of the deconvolution for all frames in the fit.
}
\examples{
plot_parameter(fit,'height')
}

