% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echem_plot.R
\name{plotechem_vertical}
\alias{plotechem_vertical}
\title{Plot electrochemisty data (vertically)}
\usage{
plotechem_vertical(data, V_range, alignment = NA, show_axes = TRUE,
  V_tick_interval = 1, V_minor_tick_interval = 0.25,
  time_tick_interval = 5, time_ticks_dps = 0, tick_size_multiplier = 1)
}
\arguments{
\item{data}{Data frame of echem data}

\item{V_range}{Range of potential (Voltage) to plot}

\item{alignment}{Alignment parameters from plotNMR2D}

\item{show_axes}{(TRUE / FALSE) Should plot axes & labels}

\item{V_tick_interval}{Tick interval for the potential axis}

\item{V_minor_tick_interval}{Minor tick interval for the potential axis}

\item{time_tick_interval}{Tick interval for the time axis}
}
\description{
This function plots in-situ echem data.
}
\examples{
plotechem_vertical(data,V_range,alignment)
}

