% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu_plot.R
\name{plotinsitu}
\alias{plotinsitu}
\title{Plot electrochemisty data (vertically)}
\usage{
plotinsitu(nmr, echem, widths = c(5, 1), ...)
}
\arguments{
\item{nmr}{Data frame of nmr data}

\item{echem}{Data frame of echem data}

\item{...}{Further parameters will be passed to \code{\link[NMR.Utils]{plotNMR2D}} or \code{\link[NMR.Utils]{plotechem_vertical}}}
}
\description{
This function plots in-situ echem data.
}
\examples{
plotinsitu(data,echem,xrange=c(1500,-500), yrange='auto', plot_offset=100000, plot.colour.ranges=list(c(210,900)), plot.colour.yranges=list(c(50000,200000)),V_range=c(0,2),alignment=align)
}

