% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_plot.R
\name{plotNMR2D}
\alias{plotNMR2D}
\title{Plot 2D NMR data}
\usage{
plotNMR2D(nmrdata, xrange, yrange, plot_offset, plot.colour = TRUE,
  plot.colour.ranges = NA, plot.colour.yranges = "auto",
  xTickInterval = 200, xMinorTickInterval = 50, y_trunc = Inf,
  col = "Black", lwd = 1, shade_under = FALSE, shade_col = "grey",
  y_trunc_x_points = c(), y_trunc_amp_div = 200,
  y_trunc_label_offset_factor = 20, y_trunc_sin_period = 5,
  y_trunc_labels = c(), y_trunc_text_col = "grey",
  y_trunc_line_col = "grey", y_trunc_lwd = 2, show_axes = TRUE,
  show_RH_Tick = TRUE)
}
\arguments{
\item{nmrdata}{Data frame of (real) NMR data (ppm, intensity1, ...)}

\item{xrange}{x range of data to plot (ppm) -- should be in form c(upper, lower) or spectrum will be reversed}

\item{yrange}{Y range of data to plot (or 'auto')}

\item{plot_offset}{offset between subsequent plots}

\item{plot.colour}{TRUE}

\item{plot.colour.ranges}{list of x ranges to colour (can be NA to cover entire range)}

\item{plot.colour.yranges}{list of y ranges to scale colours over (can be 'auto')}

\item{xTickInterval}{Tick interval for x axis}

\item{xMinorTickInterval}{Minor tick interval for x axis}

\item{y_trunc}{Truncate y values that are higher than this}

\item{col}{Line colour for plot}

\item{lwd}{Line width for plot}

\item{shade_under}{TRUE / FALSE shade under the plot}

\item{shade_col}{colour to shade plot}

\item{y_trunc_x_points}{X-ranges to draw truncation lines over}

\item{y_trunc_amp_div}{Factor to divide amplitude of truncation line}

\item{y_trunc_label_offset_factor}{Factor to offset truncation line label by}

\item{y_trunc_sin_period}{Period of sine wave for truncation line}

\item{y_trunc_labels}{List of labels for truncation line}

\item{y_trunc_text_col}{Colour of text for truncation line}

\item{y_trunc_line_col}{Colour of truncation line}

\item{y_trunc_lwd}{Line width of truncation line}

\item{show_axes}{(TRUE / FALSE) Should plot axes & labels}

\item{show_RH_Tick}{(TRUE / FALSE) Show last tick on RHS}
}
\value{
List of alignment_parameters
}
\description{
This function plots 2D NMR Data.
}
\examples{
plotNMR2D(data, xrange=c(1500,-500), yrange='auto', plot_offset=100000, plot.colour.ranges=list(c(210,900)), plot.colour.yranges=list(c(50000,200000)))
}
