% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT-fit_object.R
\name{prepare_result_data_frame}
\alias{prepare_result_data_frame}
\title{Generares the result data frame}
\usage{
prepare_result_data_frame(fit, n)
}
\arguments{
\item{fit}{The fit object to use}

\item{n}{The number of rows to add to the data frame}
}
\value{
The fit object
}
\description{
This function creates an empty data frame to hold the fitting resutls
}
\examples{
prepare_result_data_frame(fit, n)
}
\keyword{internal}
