% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_import.R
\name{process_two_column_files}
\alias{process_two_column_files}
\title{Process a folder of two column text files}
\usage{
process_two_column_files(files)
}
\arguments{
\item{files}{List of file paths to read}
}
\value{
A data frame containing the NMR data (ppm,intensity1, intensity2, ...)
}
\description{
This function attempts process 1D data files in a directory
}
\examples{
process_two_column_files(files)
}
\keyword{internal}

