% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_curves.R
\name{pseudoVoigt}
\alias{pseudoVoigt}
\title{Pseudo Voigt curve}
\usage{
pseudoVoigt(x, height, centre, hwhm, shape)
}
\arguments{
\item{x}{A vector of x coordinates}

\item{height}{The height parameter}

\item{centre}{The centre parameter}

\item{hwhm}{The hwhm parameter}

\item{shape}{The shape parameter (0--1)}
}
\value{
The value(s) of the pseudo voigt function evaluated with the specified parameters
}
\description{
This function defines the pseudo voigt curve.
}
\examples{
pseudoVoigt(10,10,0,2,0.5)
pseudoVoigt(c(1:10),10,0,2,0.5)

f=function(x) {pseudoVoigt(x,height=10,centre=0,hwhm=2,shape=0.5)}
curve(f,from=-5,to=5)
}

