% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPORT-nmr-acqu.R
\name{read.acqu}
\alias{read.acqu}
\title{Read Acquisition Parameters}
\usage{
read.acqu(dir)
}
\arguments{
\item{dir}{Path to raw data directory (up to expNo)}
}
\value{
A list of parameters
}
\description{
This function reads acquisition pararameters for NMR Data.
}
\examples{
read.acqu("/path/to/nmr/expName/expNo")
}
