% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATMC_log_import.R
\name{read.ATMC}
\alias{read.ATMC}
\title{Read ATMC log file}
\usage{
read.ATMC(log_file_path, version = 1, experimentNumber = 1)
}
\arguments{
\item{log_file_path}{Path to txt file}

\item{version}{Version of the log file (0 == orignal backup log -- no-longer used. 1 == Version prior to March 2016 upgrade.)}

\item{experimentNumber}{Index of the experiment in the current log file (currently only used if version=0)}
}
\value{
A vector of offsets giving the number of hours after the first scan that each scan was started.
}
\description{
Calculates offsets in hours from first scan based on ATMC log file
}
\examples{
read.ATMC("/path/to/file.txt")
}
