% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPORT-ATMC.R
\name{read.ATMC0}
\alias{read.ATMC0}
\title{Read ATMC (version 0 -- backup) log file}
\usage{
read.ATMC0(log_file_path, version = 1, experimentNumber = 1)
}
\arguments{
\item{log_file_path}{Path to txt file}

\item{experimentNumber}{Index of the experiment in the current log file}
}
\value{
A vector of offsets giving the number of hours after the first scan that each scan was started.
}
\description{
Calculates offsets in hours from first scan based on ATMC log file
}
\examples{
read.ATMC0("/path/to/file.txt")
}
