% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATMC_log_import.R
\name{read.ATMC1}
\alias{read.ATMC1}
\title{Read ATMC (version 1) log file}
\usage{
read.ATMC1(log_file_path, experimentNumber = 1)
}
\arguments{
\item{log_file_path}{Path to txt file}

\item{experimentNumber}{Index of the experiment in the current log file (currently not used)}
}
\value{
A vector of offsets giving the number of hours after the first scan that each scan was started.
}
\description{
Calculates offsets in hours from first scan based on ATMC log file
}
\examples{
read.ATMC1("/path/to/file.txt")
}
\keyword{internal}
