% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATMC_log_import.R
\name{read.ATMC2}
\alias{read.ATMC2}
\title{Read ATMC (version 2) log file}
\usage{
read.ATMC2(log_file_path)
}
\arguments{
\item{log_file_path}{Path to txt file}
}
\value{
A vector of offsets giving the number of hours after the first scan that each scan was started.
}
\description{
Calculates offsets in hours from first scan based on ATMC log file
}
\examples{
read.ATMC2("/path/to/file.txt")
}
