% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPORT-ATMC.R
\name{read.ATMC_generic}
\alias{read.ATMC_generic}
\title{Read ATMC log file}
\usage{
read.ATMC_generic(log_file_path, searchString)
}
\arguments{
\item{log_file_path}{Path to txt file}

\item{searchString}{String indicating the start of a new scan}
}
\value{
A vector of offsets giving the number of hours after the first scan that each scan was started.
}
\description{
Calculates offsets in hours from first scan based on ATMC log file
}
\examples{
read.ATMC_generic("/path/to/file.txt", "search_string")
}
\keyword{internal}
