% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echem_import.R
\name{read.echem}
\alias{read.echem}
\title{Read Echem Data}
\usage{
read.echem(echem_file, cycler_type)
}
\arguments{
\item{echem_file}{Path to data file}

\item{cycler_type}{Name of the cycler}
}
\value{
A data frame containing the echem data
}
\description{
This function is used to import echem data.
}
\examples{
read.echem("/path/to/file.mpt","biologic")
read.echem("/path/to/file.xls","arbin")
read.echem("/path/to/file.txt","land")
}

