% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.nmr.directory.text.R
\name{read.nmr.directory.text.1d.1col}
\alias{read.nmr.directory.text.1d.1col}
\title{Process a folder of one column text files}
\usage{
read.nmr.directory.text.1d.1col(files)
}
\arguments{
\item{files}{List of file paths to read}
}
\value{
A data frame containing the NMR data (ppm,intensity1, intensity2, ...)
}
\description{
This function attempts process 1D data files in a directory
}
\examples{
read.nmr.directory.1d.2col(files)
}
\keyword{internal}
