% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.nmr.text.1d.R
\name{read.nmr.text.1d.1col.ppm}
\alias{read.nmr.text.1d.1col.ppm}
\title{Read ppm scale for 1D data}
\usage{
read.nmr.text.1d.1col.ppm(file)
}
\arguments{
\item{file}{Path to txt file saved using topspin's File->Save->Save visible region as text file (with or without imaginary)}
}
\value{
A vector containing the ppm scale
}
\description{
This function attempts to determine the ppm scale from a bruker text file by examining the preamble.
}
\examples{
read.nmr.text.1d.1col.ppm("/path/to/file.txt")
}
\keyword{internal}
