% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.nmr.text.1d.R
\name{read.nmr.text.1d.1col}
\alias{read.nmr.text.1d.1col}
\title{Read 1D NMR Data (1 col)}
\usage{
read.nmr.text.1d.1col(nmrfile)
}
\arguments{
\item{nmrfile}{Path to txt file saved using topspin's File->Save->Save visible region as text file (with or without imaginary)}
}
\value{
A data frame containing the NMR data (ppm,intensity)
}
\description{
This function is used to import NMR data.
}
\examples{
read.nmr.text.1d.1col("/path/to/file.txt")
}
\keyword{internal}
