% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPORT-nmr.text.1d.R
\name{read.nmr.text.1d.2col}
\alias{read.nmr.text.1d.2col}
\title{Read 1D NMR Data (2 / 4 col)}
\usage{
read.nmr.text.1d.2col(file, skip, sep = "")
}
\arguments{
\item{nmrfile}{Path to txt file}
}
\value{
A data frame containing the NMR data (ppm,intensity,...)
}
\description{
This function is used to import NMR data.
}
\examples{
read.nmr.text.1d.2col("/path/to/file.txt")
}
\keyword{internal}
