% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.nmr-acqu.R
\name{read.nmr.title}
\alias{read.nmr.title}
\title{Read Title for NMR file}
\usage{
read.nmr.title(dir)
}
\arguments{
\item{dir}{Path to raw data directory (up to expNo)}
}
\value{
The title
}
\description{
This function reads the title for NMR Data.
}
\examples{
read.title("/path/to/nmr/expName/expNo")
}
