% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_import.R
\name{read.NMR1D}
\alias{read.NMR1D}
\title{Read 1D NMR Data}
\usage{
read.NMR1D(nmrfile)
}
\arguments{
\item{nmrfile}{Path to txt file saved using topspin's File->Save->Save visible region as text file (with or without imaginary)}
}
\value{
A data frame containing the NMR data (ppm,intensity)
}
\description{
This function is used to import NMR data.
}
\examples{
read.NMR1D("/path/to/file.txt")
}
\keyword{internal}
