% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_import.R
\name{read.NMR2D.raw}
\alias{read.NMR2D.raw}
\title{Read 2D NMR Data (processed, complex)}
\usage{
read.NMR2D.raw(dir, acqs)
}
\arguments{
\item{dir}{Path to raw data directory (up to expNo)}

\item{acqs}{List of acquisition parameters (\code{\link[NMR.Utils]{read.acqu}})}
}
\value{
A data frame containing the NMR data (ppm,intensity1, intensity2, ...)
}
\description{
This function is used to import NMR data.
}
\examples{
read.NMR2D.raw("/path/to/nmr/expName/expNo")
}
\keyword{internal}
