% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_import.R
\name{read.NMR2D.text}
\alias{read.NMR2D.text}
\title{Read 2D NMR Data (real only)}
\usage{
read.NMR2D.text(nmrfile, imaginary_file = NA)
}
\arguments{
\item{nmrfile}{Path to txt file saved using topspin's File->Save->Save visible region as text file (topspin doesn't support immaginary)}

\item{imaginary_file}{Path to txt file as above, but set to 90 degrees out of phase in topspin}
}
\value{
A data frame containing the NMR data (ppm,intensity1, intensity2, ...)
}
\description{
This function is used to import NMR data.
}
\examples{
read.NMR2D.text("/path/to/file.txt")
}
\keyword{internal}

