% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPORT-nmr-proc.R
\name{read.proc}
\alias{read.proc}
\title{Read procisition Parameters}
\usage{
read.proc(dir)
}
\arguments{
\item{procfile}{Path to proc(N)s file}
}
\value{
A list of parameters
}
\description{
This function reads processing pararameters for NMR Data.
}
\examples{
read.proc("/path/to/nmr/expName/expNo/pdata/1/")
}
