% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr.data.object.R
\name{read.table.nmr}
\alias{read.table.nmr}
\title{Read a table as an NMR data object}
\usage{
read.table.nmr(...)
}
\arguments{
\item{...}{parameters are passed to \code{\link{read.table}}}
}
\value{
An NMR data object containing the data
}
\description{
Read a table as an NMR data object
}
\keyword{internal}
