% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_import.R
\name{read_dir_1d_2col_as_2d}
\alias{read_dir_1d_2col_as_2d}
\title{Generate 2D data from a list of 1D data files}
\usage{
read_dir_1d_2col_as_2d(files, skip)
}
\arguments{
\item{files}{List of file paths to read}
}
\value{
A data frame containing the NMR data (ppm,intensity1, intensity2, ...)
}
\description{
This function attempts combine 1D data files into a 2D spectrum
}
\examples{
read_dir_1d_2col_as_2d(files)
}
\keyword{internal}
