% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_import.R
\name{read_ppm_1D}
\alias{read_ppm_1D}
\title{Read ppm scale for 1D data}
\usage{
read_ppm_1D(file)
}
\arguments{
\item{file}{Path to txt file saved using topspin's File->Save->Save visible region as text file (with or without imaginary)}
}
\value{
A vector containing the ppm scale
}
\description{
This function attempts to determine the ppm scale from a bruker text file by examining the preamble.
}
\examples{
read_ppm_1D("/path/to/file.txt")
}
\keyword{internal}

