% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCESS-rebin.R
\name{rebin}
\alias{rebin}
\title{Rebins the data to ensure a minimum offset between scans}
\usage{
rebin(data, min_offset, method = c("discard", "average_before",
  "average_after"))
}
\arguments{
\item{data}{2D NMR data}

\item{min_offset}{Minimum offset between adjacent scans. Scans too close together will be dealt
with according to the \code{method} option.}

\item{method}{One of the following:
\describe{
 \item{discard}{Scans that are not sufficiently offset from the previous scan are discarded.}
 \item{average_before}{Scans that are not sufficiently offset from the previous scan are averaged with the previous scan.}
 \item{average_after}{Scans that are not sufficiently offset from the previous scan are averaged with the next scan.}
}}
}
\description{
Rebins the data to ensure a minimum offset between scans
}
\examples{
rebin(data, 10)
rebin(data, 10, "average_before")
}
