% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_process.R
\name{reduceScans}
\alias{reduceScans}
\title{Reduces the number of scans in the data frame}
\usage{
reduceScans(spectra, scan_range)
}
\arguments{
\item{spectra}{A dataframe of NMR data}

\item{scan_range}{A vector of scans to include c(first, last)}
}
\value{
An idential data frame whose scans are limited to those in scan_range
}
\description{
Reduces the number of scans in the data frame
}
\examples{
reduceScans(spectra,c(1,50))
}
