% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-PLOTTING-SUBPLOTS.R
\name{register_insitu_subplot_module}
\alias{register_insitu_subplot_module}
\title{Register an interface for drawing a subplot within \code{\link{interactivePlotting}}}
\usage{
register_insitu_subplot_module(name, uiFunction, serverFunction)
}
\arguments{
\item{name}{Name of the plot type}

\item{uiFunction}{Shiny module UI function that takes an id parameter and returns a \code{\link[shiny]{tagList}}.}

\item{serverFunction}{Shiny module server function that takes \code{data} and \code{data_name} \code{\link[shiny]{reactive}} parameters and returns a \code{\link[base]{list}} with items: \cr
\describe{
 \item{plotFunc}{Function to draw the plot. Must be compatible with \code{\link{layout}}.}
 \item{script}{\code{\link[shiny]{reactive}} that returns a string containing the code to draw the plot}
}}
}
\description{
Note that this function is not persistent. You must re-register in every R session you wish to use this plot type.
}
\seealso{
\link{http://shiny.rstudio.com/articles/modules.htm}
}
