% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-PLOTTING-ANNOTATION-AAAA.R
\name{register_plot_annotation_module}
\alias{register_plot_annotation_module}
\title{Register an interface for drawing an annotation within \code{\link{interactivePlotting}}}
\usage{
register_plot_annotation_module(name, uiFunction, serverFunction)
}
\arguments{
\item{name}{Name of the annotation type}

\item{uiFunction}{Shiny module UI function that takes an id parameter and returns a \code{\link[shiny]{tagList}}.}

\item{serverFunction}{Shiny module server function that returns a \code{\link[base]{list}} with items: \cr
\describe{
 \item{arguments}{Function that returns a list of arguments to be passted to \code{drawFunc}}
 \item{drawFunc}{Function that draws the annotation. Takes a list of arguments as returned above, plus x and y coordinates as the first two arguments, and (possibly) a named xpd argument. Should NOT be reactive nor depend on the UI.}
 \item{script}{Function that returns a string containing the code to draw the annotation. Takes the same arguments as \code{drawFunc}. Should NOT be reactive nor depend on the UI.}
 \item{getBoundingRect}{Function that returns vector containing the bounding rect of the annotation for hit testing. Takes the same arguments as \code{drawFunc}, returns c(left, bottom, right, top).}
 \item{focus}{Function that highlights an annotation (typically by drawing a red border). Takes the same arguments as \code{drawFunc}.}
 \item{updateInputs}{Function that updates the UI to reflect the argument list. Takes the same arguments returned by \code{arguments}.}
}}
}
\description{
Note that this function is not persistent. You must re-register in every R session you wish to use this annotation type.
}
\seealso{
\link{http://shiny.rstudio.com/articles/modules.htm}
}
