% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_plot.R
\name{rescale.nmr}
\alias{rescale.nmr}
\title{Rescale Data}
\usage{
rescale.nmr(x, newrange)
}
\description{
This function is a workaround for an issue with \code{\link[plotrix]{rescale}}.
}
\keyword{internal}

