% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT-leastSquares.R
\name{run_fit_for_data}
\alias{run_fit_for_data}
\title{Run a fit}
\usage{
run_fit_for_data(fit, data)
}
\arguments{
\item{fit}{The fit object to use}

\item{data}{The data to fit}
}
\value{
The fit object
}
\description{
This function fits the data using the models defined previously.
}
\examples{
run_fit_for_data(fit, data)
}
