% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT-Animations.R
\name{save_animation}
\alias{save_animation}
\title{Make an animation}
\usage{
save_animation(data, fit = NULL, xlim = NULL, V_range = NULL,
  ylim = NULL, data_col = "black", data_lwd = 1,
  cols = c("forestgreen", "blue", "magenta", "cyan", "orange", "brown"),
  lwd = 1, lty = "dashed", fit_col = "red", fit_lwd = 2,
  delay = 0.1, width = 600, height = 400, type = "gif",
  name = "Animation.gif", image_dev = "png", image_type = "png", ...)
}
\arguments{
\item{data}{The insitu data object}

\item{fit}{The fit object}

\item{xlim}{x range of data to plot}

\item{ylim}{Y range of data to plot}

\item{data_col}{Colour for the data}

\item{data_lwd}{Line width for the data}

\item{cols}{Colour for each deconvolution}

\item{lwd}{Line width for each deconvolution}

\item{lty}{Line type for each deconvolution}

\item{fit_col}{Colour for the fit}

\item{fit_lwd}{Line width for the fit}

\item{delay}{Delay between frames in the animation}

\item{width}{Width of the animation}

\item{height}{Height of the animation}

\item{type}{Type of animation (gif or HTML) gif requires additional software (see \code{\link[animation]{saveGIF}})}

\item{name}{File name to output}

\item{image_dev}{Image device to use (if you get error messages about png, try setting this to 'jpeg')}

\item{image_type}{Image type to use (if you get error messages about png, try setting this to 'jpg')}

\item{...}{Additional parameters passed to the NMR plotting function}
}
\value{
None
}
\description{
This function makes an animation for the fit.
}
\examples{
save_animation(data, fit)
}
