% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{save_fit}
\alias{save_fit}
\title{Save fit}
\usage{
save_fit(fit, filename)
}
\arguments{
\item{fit}{The fit object}

\item{filename}{The path to the file (ending in .R)}
}
\value{
None
}
\description{
This function is used to save the fit as an R binary file. The resulting file can be read using (\code{\link[NMR.Utils]{load_fit}}).
}
\examples{
save_fit(fit,"/path/to/file.R")
}

