% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echem_library_integration.R
\name{setup_NMR_echem_integration}
\alias{setup_NMR_echem_integration}
\title{Allow echem library to plot NMR data}
\usage{
setup_NMR_echem_integration()
}
\description{
This function is used to integrate NMR plotting into the echem library.
}
