% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_process.R
\name{storeOffsets}
\alias{storeOffsets}
\title{Stores the time offsets in the data frame}
\usage{
storeOffsets(spectra, offsets)
}
\arguments{
\item{spectra}{A dataframe of NMR data}

\item{offsets}{A vector of offsets}
}
\value{
An idential data frame whose names correspond to the offsets
}
\description{
Stores the time offsets in the data frame
}
\examples{
storeOffsets(data,offsets)
}

