% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_remaining.R
\name{time_remaining_print}
\alias{time_remaining_print}
\title{Iterate a time remaining indicator}
\usage{
time_remaining_print(current_scan = NULL, envir = NULL)
}
\arguments{
\item{current_scan}{The current scan number.}
}
\value{
The environment containing the timing parameters.
}
\description{
This function iterates a time remaining indicator. It should not be called manually, rather the function returned from \code{\link[NMR.Utils]{time_remaining_start}} should be used.
}
\examples{
time_remaining_start(10)
f=time_remaining_start(10)
for(i in 1:10) {
  Sys.sleep(1)
  f(i)
  Sys.sleep(1)
  f()
}
}
\keyword{internal}

