% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_remaining.R
\name{time_remaining_start}
\alias{time_remaining_start}
\title{Start a time remaining indicator}
\usage{
time_remaining_start(max_scan)
}
\arguments{
\item{max_scan}{The number of scans that will be performed during the time the indicator is shown.}
}
\value{
A function to be called with the current scan number
}
\description{
This function starts a time remaining indicator.
}
\examples{
time_remaining_start(10)
f=time_remaining_start(10)
for(i in 1:10) {
  Sys.sleep(1)
  f(i)
  Sys.sleep(1)
  f()
}
}
\keyword{internal}
