% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echem_import.R
\name{vswitch}
\alias{vswitch}
\title{Implementation of switch function in R}
\usage{
vswitch(EXPR, ...)
}
\arguments{
\item{EXPR}{The vector to apply the changes to}

\item{...}{The changes to apply}
}
\value{
A vector with the changes applied
}
\description{
This function is used internally to ease the importing of biologic data from different versions of ec-lab.
}
\examples{
vswitch(input, changes ...)
}
\keyword{internal}

