% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_jms.reexports.R
\name{reexports}
\alias{reexports}
\alias{iplot}
\alias{export}
\alias{xlab}
\alias{ylab}
\alias{y2lab}
\alias{xlab<-}
\alias{ylab<-}
\alias{y2lab<-}
\alias{xcol}
\alias{ycol}
\alias{y2col}
\alias{xcol<-}
\alias{ycol<-}
\alias{y2col<-}
\alias{xscale}
\alias{yscale}
\alias{y2scale}
\alias{xscale<-}
\alias{yscale<-}
\alias{y2scale<-}
\alias{make_background}
\alias{expand_baseline_parameters}
\alias{make_backgrounds}
\alias{undersample}
\title{Reexported functions from jms.classes}
\description{
Reexported functions from jms.classes
}
\keyword{internal}
