% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ex_situ.R
\name{plot.nmr.data.object}
\alias{plot.nmr.data.object}
\title{Plot ex situ NMR data}
\usage{
\method{plot}{nmr.data.object}(data, ...)
}
\arguments{
\item{data}{An nmr.data.object}

\item{...}{Additional parameters passed to \code{\link{plot_ex_situ_nmr}}}
}
\description{
This function is used to plot ex situ NMR data.
}
