% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu_plot.R
\name{plot.nmr2dinsitu.data.object}
\alias{plot.nmr2dinsitu.data.object}
\title{Plot In situ NMR data}
\usage{
\method{plot}{nmr2dinsitu.data.object}(nmr, widths = c(3, 1), ...)
}
\arguments{
\item{nmr}{In situ nmr data}

\item{widths}{Ratio of NMR to echem plot sizes}

\item{...}{Further parameters will be passed to \code{\link[NMR.Utils]{plot.nmr2d.data.object}} or \code{\link[NMR.Utils]{plot_echem_vertical}}}
}
\description{
This function plots in-situ NMR data.
}
\examples{
plot(data)
}
