% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT-Animations.R
\name{plot_animation_frame}
\alias{plot_animation_frame}
\title{Plot a raw frame for a fit animation}
\usage{
plot_animation_frame(data, frame_number, fit = NULL, xlim = NULL,
  V_range = NULL, ylim = NULL, data_col = "black", data_lwd = 1,
  cols = c("forestgreen", "blue", "magenta", "cyan", "orange", "brown"),
  lwd = 1, lty = "dashed", fit_col = "red", fit_lwd = 2, ...)
}
\arguments{
\item{data}{The insitu data object}

\item{frame_number}{Which frame to plot}

\item{fit}{The fit object}

\item{xlim}{x range of data to plot}

\item{ylim}{Y range of data to plot}

\item{data_col}{Colour for the data}

\item{data_lwd}{Line width for the data}

\item{cols}{Colour for each deconvolution}

\item{lwd}{Line width for each deconvolution}

\item{lty}{Line type for each deconvolution}

\item{fit_col}{Colour for the fit}

\item{fit_lwd}{Line width for the fit}

\item{...}{Additional parameters passed to the NMR plotting function}
}
\value{
None
}
\description{
This function plots a single frame from the fit.
}
\examples{
plot_animation_frame(data, 100, fit)
}
\keyword{internal}
