% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu_plot.R
\name{plot_echem_vertical}
\alias{plot_echem_vertical}
\title{Plot electrochemisty data (vertically)}
\usage{
plot_echem_vertical(data, V_range = NA, time_start = 0, time_end = Inf,
  yrange = c(0, 1), offset_start = 0, offset_end = 1, show_axes = TRUE,
  V_tick_interval = NA, V_minor_tick_interval = NA,
  time_tick_interval = NA, time_ticks_dps = 0, tick_size_multiplier = 1,
  xaxismline = -0.8, xaxislabelmline = 1.1)
}
\arguments{
\item{data}{Echem data}

\item{V_range}{Voltage range of echem data to plot}

\item{time_start, time_end}{Range of time to plot}

\item{offset_start, offset_end}{y values of 1st and last data point}

\item{show_axes}{--ignored--}

\item{V_tick_interval, V_minor_tick_interval, V_minor_tick_interval}{--ignored--}

\item{time_tick_interval, time_ticks_dps, tick_size_multiplier}{--ignored--}

\item{xaxismline}{Margin line on which to draw the x axis}

\item{xaxislabelmline}{Margin line on which to draw the x axis title}

\item{y_range}{Y range to scale the data to}
}
\description{
This function plots in-situ echem data.
}
