% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT-nmr2dinsitu.data.object.R
\name{plot_echem_vertical}
\alias{plot_echem_vertical}
\title{Plot electrochemisty data (vertically)}
\usage{
plot_echem_vertical(data, V_range = NULL, ylim = NULL,
  echemAxes = c(1, 4), echemDiv = c(2, 5), xaxismline = -0.8,
  xaxislabelmline = 1.1, yaxismline = -0.8, yaxislabelmline = 1.1,
  upperVTickLimit = NA, lowerVTickLimit = NA, ticklabels = c(T, T,
  T), ticksOut = c(T, T), forcedVInterval = NA, forcePrint = FALSE,
  centreTimeTitle = T, ...)
}
\arguments{
\item{data}{Echem data}

\item{V_range}{X range of echem data to plot}

\item{ylim}{Numeric vectors of length 2, giving the x, y (and y2) coordinate ranges.}

\item{echemAxes}{Axes for the echem plot}

\item{echemDiv}{Number of minor ticks per major c(voltage, y)}

\item{xaxismline, yaxismline}{Margin line on which to draw the axes}

\item{xaxislabelmline, yaxislabelmline}{Margin line on which to draw the axis titles}

\item{upperVTickLimit, lowerVTickLimit}{Voltage tick limits}

\item{ticklabels}{Print labels?}

\item{ticksOut}{Ticks point outwards?}

\item{forcedVInterval}{Forced voltage tick interval}

\item{forcePrint}{Attempt to print all labels even if they are closely spaced}

\item{centreTimeTitle}{Centre the y axis label to the tick marks?}

\item{...}{Additional parameters are passed to the underlying function calls}
}
\description{
This function plots in-situ echem data. It is effectively the same as plot,
but swaps the x and y axes, and renames some arguments to avoid clashes with
\code{\link{plot.nmr2d.data.object}}.
}
