% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPORT-ATMC.R
\name{read.ATMC1}
\alias{read.ATMC1}
\title{Read ATMC (version 1) log file}
\usage{
read.ATMC1(log_file_path)
}
\arguments{
\item{log_file_path}{Path to txt file}
}
\value{
A vector of offsets giving the number of hours after the first scan that each scan was started.
}
\description{
Calculates offsets in hours from first scan based on ATMC log file
}
\examples{
read.ATMC1("/path/to/file.txt")
}
