% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT-Animations.R
\name{save_animation}
\alias{save_animation}
\title{Make an animation}
\usage{
save_animation(fit, data, echem, xrange, V_range, yrange = "auto",
  include_fit = TRUE, data_col = "black", data_lwd = 1,
  cols = c("forestgreen", "blue", "magenta", "cyan", "orange", "brown"),
  lwd = 1, lty = "dashed", fit_col = "red", fit_lwd = 2,
  xTickInterval = 50, delay = 0.1, width = 400, height = 400,
  type = "gif", name = "Animation.gif", image_dev = "png",
  image_type = "png")
}
\arguments{
\item{fit}{The fit object}

\item{data}{The data object}

\item{echem}{The echem object}

\item{xrange}{x range of data to plot}

\item{yrange}{Y range of data to plot (or 'auto')}

\item{include_fit}{include the deconvolution etc. in the plot}

\item{data_col}{Colour for the data}

\item{data_lwd}{Line width for the data}

\item{cols}{Colour for each deconvolution}

\item{lwd}{Line width for each deconvolution}

\item{lty}{Line type for each deconvolution}

\item{fit_col}{Colour for the fit}

\item{fit_lwd}{Line width for the fit}

\item{xTickInterval}{Tick interval for x axis}

\item{delay}{Delay between frames in the animation}

\item{width}{Width of the animation}

\item{height}{Height of the animation}

\item{type}{Type of animation (gif or HTML) gif requires additional software (see \code{\link[animation]{saveGIF}})}

\item{name}{File name to output}

\item{image_dev}{Image device to use (if you get error messages about png, try setting this to 'jpeg')}

\item{image_type}{Image type to use (if you get error messages about png, try setting this to 'jpg')}
}
\value{
None
}
\description{
This function makes an animation for the fit.
}
\examples{
save_animation(fit,data,echem,c(200,-200),c(0,2))
}
